refactor(auth): canonicalize auth_v2 package structure
