+++ backend/core/auth_v2/__init__.py
@@ -1,4 +1,15 @@
 from . import models
-from .router import router
+from .api import oauth, router
+from .repositories import AuthRepositoryV2
+from .security import Argon2Hasher, RefreshTokenManager
+from .services import AuthService
 
-__all__ = ["models", "router"]
+__all__ = [
+    "Argon2Hasher",
+    "AuthRepositoryV2",
+    "AuthService",
+    "RefreshTokenManager",
+    "models",
+    "oauth",
+    "router",
+]
