+++ backend/core/auth_v2/api/oauth_client.py
@@ -1,0 +1,16 @@
+from authlib.integrations.starlette_client import OAuth
+
+from core.config import settings
+
+oauth = OAuth()
+oauth.register(
+    name="google",
+    client_id=settings.google_oauth_client_id,
+    client_secret=settings.google_oauth_client_secret,
+    server_metadata_url="https://accounts.google.com/.well-known/openid-configuration",
+    client_kwargs={
+        "scope": "openid email profile",
+    },
+)
+
+__all__ = ["oauth"]
