+++ backend/core/auth_v2/api/router.py
@@ -1,44 +1,25 @@
 import uuid
 from typing import Annotated
 
-from authlib.integrations.starlette_client import OAuth
 from fastapi import APIRouter, Cookie, Depends, HTTPException, Request, Response, status
 from fastapi.responses import RedirectResponse
 from loguru import logger
 
 from core.config import settings
 
-from .api import get_auth_service, get_current_user_id_v2, get_request_client_context
-from .config import (
-    ACCESS_TOKEN_COOKIE_HTTPONLY,
-    ACCESS_TOKEN_COOKIE_NAME,
-    ACCESS_TOKEN_COOKIE_PATH,
-    ACCESS_TOKEN_COOKIE_SAMESITE,
-    ACCESS_TOKEN_COOKIE_SECURE,
-    REFRESH_TOKEN_COOKIE_HTTPONLY,
-    REFRESH_TOKEN_COOKIE_NAME,
-    REFRESH_TOKEN_COOKIE_PATH,
-    REFRESH_TOKEN_COOKIE_SAMESITE,
-    REFRESH_TOKEN_COOKIE_SECURE,
-    REFRESH_TOKEN_TTL_SECONDS,
+from ..config import REFRESH_TOKEN_COOKIE_NAME
+from ..exceptions import InvalidRefreshTokenError, UserNotFoundError
+from ..schemas import UserResponse
+from ..services import AuthService
+from .cookies import clear_auth_cookies, set_auth_cookies
+from .dependencies import (
+    get_auth_service,
+    get_current_user_id_v2,
+    get_request_client_context,
 )
-from .exceptions import InvalidRefreshTokenError, UserNotFoundError
-from .schemas import UserResponse
-from .service import AuthService
+from .oauth_client import oauth
 
 router = APIRouter(prefix="/auth", tags=["auth", "google-auth"])
-
-
-oauth = OAuth()
-oauth.register(
-    name="google",
-    client_id=settings.google_oauth_client_id,
-    client_secret=settings.google_oauth_client_secret,
-    server_metadata_url="https://accounts.google.com/.well-known/openid-configuration",
-    client_kwargs={
-        "scope": "openid email profile",
-    },
-)
 
 
 def _frontend_auth_redirect(path: str) -> str:
@@ -46,62 +27,16 @@ def _frontend_auth_redirect(path: str) -> str:
     return f"{frontend_base_url.rstrip('/')}{path}"
 
 
-def _set_auth_cookies(
-    response: Response,
-    *,
-    access_token: str,
-    refresh_token: str,
-) -> None:
-    response.set_cookie(
-        key=ACCESS_TOKEN_COOKIE_NAME,
-        value=access_token,
-        httponly=ACCESS_TOKEN_COOKIE_HTTPONLY,
-        secure=ACCESS_TOKEN_COOKIE_SECURE,
-        samesite=ACCESS_TOKEN_COOKIE_SAMESITE,
-        max_age=None,
-        path=ACCESS_TOKEN_COOKIE_PATH,
-    )
-    response.set_cookie(
-        key=REFRESH_TOKEN_COOKIE_NAME,
-        value=refresh_token,
-        httponly=REFRESH_TOKEN_COOKIE_HTTPONLY,
-        secure=REFRESH_TOKEN_COOKIE_SECURE,
-        samesite=REFRESH_TOKEN_COOKIE_SAMESITE,
-        max_age=REFRESH_TOKEN_TTL_SECONDS,
-        path=REFRESH_TOKEN_COOKIE_PATH,
-    )
-
-
-def _clear_auth_cookies(response: Response) -> None:
-    response.delete_cookie(
-        key=ACCESS_TOKEN_COOKIE_NAME,
-        path=ACCESS_TOKEN_COOKIE_PATH,
-        httponly=ACCESS_TOKEN_COOKIE_HTTPONLY,
-        secure=ACCESS_TOKEN_COOKIE_SECURE,
-        samesite=ACCESS_TOKEN_COOKIE_SAMESITE,
-    )
-    response.delete_cookie(
-        key=REFRESH_TOKEN_COOKIE_NAME,
-        path=REFRESH_TOKEN_COOKIE_PATH,
-        httponly=REFRESH_TOKEN_COOKIE_HTTPONLY,
-        secure=REFRESH_TOKEN_COOKIE_SECURE,
-        samesite=REFRESH_TOKEN_COOKIE_SAMESITE,
-    )
-
-
 @router.get("/login")
 async def login(request: Request) -> RedirectResponse:
     google = oauth.create_client("google")
     redirect_uri = str(request.url_for("callback"))
-    # The callback URL is the endpoint that Google will redirect the user to after they complete authentication.
-    # Here, request.url_for("callback") dynamically generates the absolute URL for the "/callback" route,
-    # ensuring that OAuth will return the authenticated user to the correct handler in this FastAPI app.
     logger.info(f"Redirecting to Google OAuth: {redirect_uri}")
     return await google.authorize_redirect(  # type: ignore[no-any-return]
         request,
         redirect_uri,
-        access_type="offline",  # request refresh token (for future Google API use)
-        prompt="consent",  # ensure refresh token is returned
+        access_type="offline",
+        prompt="consent",
     )
 
 
@@ -123,14 +58,14 @@ async def callback(
             ip=ip,
         )
         response = RedirectResponse(url=_frontend_auth_redirect("/auth/success"))
-        _set_auth_cookies(
+        set_auth_cookies(
             response,
             access_token=result.tokens.access_token,
             refresh_token=result.tokens.refresh_token,
         )
         return response
-    except Exception as e:
-        logger.error(f"Error authorizing access token: {e}")
+    except Exception as exc:
+        logger.error(f"Error authorizing access token: {exc}")
         return RedirectResponse(url=_frontend_auth_redirect("/auth/failure"))
 
 
@@ -169,7 +104,7 @@ async def refresh(
             status_code=status.HTTP_401_UNAUTHORIZED,
             detail=str(exc),
         )
-    _set_auth_cookies(
+    set_auth_cookies(
         response,
         access_token=tokens.access_token,
         refresh_token=tokens.refresh_token,
@@ -187,6 +122,6 @@ async def logout(
     ] = None,
 ) -> Response:
     await service.logout(refresh_token)
-    _clear_auth_cookies(response)
+    clear_auth_cookies(response)
     response.status_code = status.HTTP_204_NO_CONTENT
     return response
