+++ backend/core/auth_v2/repositories/user.py
@@ -1,0 +1,23 @@
+import uuid
+
+from sqlalchemy import select
+from sqlalchemy.ext.asyncio import AsyncSession
+
+from ..models import User
+
+
+class UserRepository:
+    def __init__(self, db: AsyncSession):
+        self.db = db
+
+    async def get_user_by_id(self, user_id: uuid.UUID) -> User | None:
+        return await self.db.get(User, user_id)
+
+    async def get_user_by_email(self, email: str) -> User | None:
+        query = select(User).where(User.email == email)
+        result = await self.db.execute(query)
+        return result.scalar_one_or_none()
+
+    async def create_user(self, user: User) -> User:
+        self.db.add(user)
+        return user
