+++ backend/core/auth_v2/services/__init__.py
@@ -1,0 +1,11 @@
+from .auth import AuthService, CallbackResult
+from .oauth import OAuthService
+from .session import AuthTokens, SessionService
+
+__all__ = [
+    "AuthService",
+    "AuthTokens",
+    "CallbackResult",
+    "OAuthService",
+    "SessionService",
+]
