+++ backend/tests/test_auth_v2_repository.py
@@ -3,8 +3,8 @@ from datetime import datetime, timedelta, timezone
 from sqlalchemy.ext.asyncio import AsyncSession
 
 from core.auth.models.user import User
+from core.auth_v2 import AuthRepositoryV2
 from core.auth_v2.models import GoogleOAuthAccount
-from core.auth_v2.repository import AuthRepositoryV2
 
 
 async def test_create_google_oauth_account_links_to_user(
